requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Authentication and Security
python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration and Validation
pydantic-settings==2.1.0
email-validator==2.1.0

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3

# File Operations
# aiofiles==23.2.1

spleeter==2.4.0

pydub==0.25.1
fastdtw==0.3.4
# scipy==1.11.4

# Testing (development only)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# httpx==0.25.2